+++ run_tests.py
@@ -3,7 +3,7 @@
 import sys
 import unittest
 
-import PyTerminalCommander.tests
+import PyTerminalCommander.tests.super_move
 
 res = unittest.TextTestRunner().run(PyTerminalCommander.tests.Tests)
 sys.exit(len(res.failures))
